Add test instructions and fix region disk test
As a step toward adding integration testing into
the CI process for this project, adding instructions
on how to run the tests.

compute_region_disk was used as a test. Fixed
the tests which was previously hard-coded to
the `graphite-playground` GCP project to work
with the project in the user-configure .ini file.