disable spanner_database_info tests, python to 3.9
The client's List continues to be flakey. disabling until
the endpoint is stable.

Switching tests to use ansible 2.14, as 2.13 has issues with
spawning a docker container.

Upgrading the base python version as a consequence.

Removing the need for docker from unit tests, since that was failing
and also non-critical to run unit tests.