test: disable gcp_bigtable_instance test
The test is still flakey currently. A few attempts have
been made to fix it, but it seems that the bigtable API
does not consistently return the resource after is has
been created.

While exploring ways to make it more robust, disabling the
test to keep the CI tests from being unreliable to the point
of useless.