chore: pin ansible-lint to 6.13.1
ansible-lint was updated and the GitHub
action did not pin the dependency, resulting
in the repository to fail sanity tests.

Updating the repository to adhere to new fatal
linter rules, but also pinning the linter to
prevent failures that may be unrelated to the
particular commit.

Updating usages for python3.8 to 3.9 as ansible-lint
is dropping support for 3.8.