Separate request and response
Makes sense to keep in separate dictionaries, not sure why I didn't do
this from the start...

Had to rewrite a lot of this

Because it turns out that new-style APIs like to change the
update/create/delete verbs, I had to make this more flexible by
referencing the operations from the Resource itself.
Wrote a configuration class just to avoid using paren notation for
dictionaries, and because it may require more massaging in the future.

if_object should return a nested dict

O() is for option names

Add some exclusions to dot_fields

labels, annotations and tags are free-form dictionaries, it makes sense
to expose their top-level, but makes no sense to expose their members